requirements.txt
langchain
faiss-cpu
openai
tiktoken
PyMuPDF
python-dotenv
